Speed up library scan
* Workaround for SQLite limitation
   Year can't be null so it is -1 instead.
* Check for null of strings that can't be null when scanning
* Add BaseNameStringList
* Rename ArtistsOfTrack and GenresOfTrack.
   The purpose of these is now a bit more generic.
* Remove ITagCreator. NSubstitute doesn't need it.

+++ LalaDb/Data/LalaTags.cs
@@ -28,7 +28,7 @@ namespace LalaDb.Data
             if (track!.GenreTracks is { }) { builder = builder.WithGenre(GenreTackListToStringList(track?.GenreTracks)); }
             if (track!.TrackId is { }) { builder = builder.WithTrackId(track.TrackId); }
             if (track!.TrackNumber is { }) { builder = builder.WithTrackNumber((uint)track.TrackNumber); }
-            if (track!.Year is { }) { builder = builder.WithYear((uint)track.Year); }
+            if (track!.Year is { }) { builder = builder.WithYear(track.Year); }
             var tags = builder.Build();
 
             return new LalaTags(tags);
